refactor: 매칭 기능 입력에 대한 비즈니스 로직 Service 클래스 추가 및 이동
