+++ src/main/java/pairmatching/domain/Course.java
@@ -1,4 +1,4 @@
-package pairmatching.model;
+package pairmatching.domain;
 
 import java.util.Arrays;
 import java.util.List;
