+++ src/main/java/pairmatching/domain/SelectMissionDto.java
@@ -1,4 +1,4 @@
-package pairmatching.model;
+package pairmatching.domain;
 
 public class SelectMissionDto {
     Course course;
